// core/admin-dashboard/store/category/action-types.ts
export enum ActionTypes {
  GET_CATEGORIES = "GET_CATEGORIES",
  GET_CATEGORY = "GET_CATEGORY",
  CREATE_CATEGORY = "CREATE_CATEGORY",
  UPDATE_CATEGORY = "UPDATE_CATEGORY",
  DELETE_CATEGORY = "DELETE_CATEGORY",
  UPLOAD_CATEGORY_THUMBNAIL = "UPLOAD_CATEGORY_THUMBNAIL",
  HARD_DELETE_CATEGORY = "HARD_DELETE_CATEGORY",
}