// core/admin-dashboard/store/category/mutation-types.ts
export enum MutationTypes {
  SET_CATEGORY = "SET_CATEGORY",
  SET_CATEGORIES = "SET_CATEGORIES",
  UPDATE_CATEGORY_DATA = "UPDATE_CATEGORY_DATA",
  SET_CATEGORY_ANALYS_DATA = "SET_CATEGORY_ANALYS_DATA",
  SET_CATEGORY_PAGINATION = "SET_CATEGORY_PAGINATION",
  UPDATE_CATEGORY_PAGINATION = "UPDATE_CATEGORY_PAGINATION",
}